fix: execulta alteracao apenas em ambiente que nao seja teste

+++ database/migrations/2021_06_22_230316_drop_foreign_user_in_pacientes_table.php
@@ -13,13 +13,16 @@ class DropForeignUserInPacientesTable extends Migration
      * @return void
      */
     public function up()
-    {        
-        Schema::table('pacientes', function (Blueprint $table) {
-            $table->dropForeign('pacientes_user_id_foreign');
-            $table->dropColumn('user_id');
-        });
+    {   
+        if (!app()->environment('testing'))
+        {
+            Schema::table('pacientes', function (Blueprint $table) {
+                $table->dropForeign('pacientes_user_id_foreign');
+                $table->dropColumn('user_id');
+            });
 
-        DB::table('users')->where('role', '=', 'paciente')->delete();
+            DB::table('users')->where('role', '=', 'paciente')->delete();
+        }
     }
 
     /**
